Move injected context fragments into core
