+++ codex-rs/core/src/context/mod.rs
@@ -1,12 +1,28 @@
+mod approved_command_prefix_saved;
+mod collaboration_mode_instructions;
 mod contextual_user_message;
 mod environment_context;
 mod fragment;
+mod guardian_followup_review_reminder;
+mod hook_additional_context;
+mod image_generation_instructions;
+mod model_switch_instructions;
+mod network_rule_saved;
+mod permissions_instructions;
+mod personality_spec_instructions;
+mod plugin_instructions;
+mod realtime_end_instructions;
+mod realtime_start_instructions;
+mod realtime_start_with_instructions;
 mod skill_instructions;
+mod spawn_agent_instructions;
 mod subagent_notification;
 mod turn_aborted;
 mod user_instructions;
 mod user_shell_command;
 
+pub(crate) use approved_command_prefix_saved::ApprovedCommandPrefixSaved;
+pub(crate) use collaboration_mode_instructions::CollaborationModeInstructions;
 pub(crate) use contextual_user_message::is_contextual_user_fragment;
 pub(crate) use contextual_user_message::is_memory_excluded_contextual_user_fragment;
 pub(crate) use contextual_user_message::parse_visible_hook_prompt_message;
@@ -14,7 +30,19 @@ pub(crate) use environment_context::EnvironmentContext;
 pub(crate) use fragment::ContextualUserFragment;
 pub(crate) use fragment::FragmentRegistration;
 pub(crate) use fragment::FragmentRegistrationProxy;
+pub(crate) use guardian_followup_review_reminder::GuardianFollowupReviewReminder;
+pub(crate) use hook_additional_context::HookAdditionalContext;
+pub(crate) use image_generation_instructions::ImageGenerationInstructions;
+pub(crate) use model_switch_instructions::ModelSwitchInstructions;
+pub(crate) use network_rule_saved::NetworkRuleSaved;
+pub(crate) use permissions_instructions::PermissionsInstructions;
+pub(crate) use personality_spec_instructions::PersonalitySpecInstructions;
+pub(crate) use plugin_instructions::PluginInstructions;
+pub(crate) use realtime_end_instructions::RealtimeEndInstructions;
+pub(crate) use realtime_start_instructions::RealtimeStartInstructions;
+pub(crate) use realtime_start_with_instructions::RealtimeStartWithInstructions;
 pub(crate) use skill_instructions::SkillInstructions;
+pub(crate) use spawn_agent_instructions::SpawnAgentInstructions;
 pub(crate) use subagent_notification::SubagentNotification;
 pub(crate) use turn_aborted::TurnAborted;
 pub(crate) use user_instructions::UserInstructions;
