+++ codex-rs/core/src/context/prompts/permissions/approval_policy/never.md
