+++ codex-rs/core/src/context/prompts/permissions/approval_policy/on_failure.md
