+++ codex-rs/core/src/context/prompts/permissions/approval_policy/on_request.md
