+++ codex-rs/core/src/context/prompts/permissions/approval_policy/on_request_rule_request_permission.md
