+++ codex-rs/core/src/context/prompts/permissions/approval_policy/unless_trusted.md
