+++ codex-rs/core/src/context/prompts/permissions/sandbox_mode/danger_full_access.md
