+++ codex-rs/core/src/context/prompts/permissions/sandbox_mode/read_only.md
