+++ codex-rs/core/src/context/prompts/permissions/sandbox_mode/workspace_write.md
