+++ codex-rs/core/src/context/prompts/realtime/realtime_start.md
