+++ codex-rs/core/src/plugins/injection.rs
@@ -2,10 +2,11 @@ use std::collections::BTreeSet;
 use std::collections::HashMap;
 
 use codex_connectors::metadata::connector_display_label;
-use codex_protocol::models::DeveloperInstructions;
 use codex_protocol::models::ResponseItem;
 
 use crate::connectors;
+use crate::context::ContextualUserFragment;
+use crate::context::PluginInstructions;
 use crate::plugins::PluginCapabilitySummary;
 use crate::plugins::render_explicit_plugin_instructions;
 use codex_mcp::CODEX_APPS_MCP_SERVER_NAME;
@@ -52,8 +53,8 @@ pub(crate) fn build_plugin_injections(
                 .into_iter()
                 .collect::<Vec<_>>();
             render_explicit_plugin_instructions(plugin, &available_mcp_servers, &available_apps)
-                .map(DeveloperInstructions::new)
-                .map(ResponseItem::from)
+                .map(PluginInstructions::new)
+                .map(ContextualUserFragment::into)
         })
         .collect()
 }
